+++ Application/Interface/IAgendamentoAppService.cs
@@ -8,7 +8,7 @@ namespace Application.Interface
     public interface IAgendamentoAppService
     {
         Task<ResponseAgendamentoViewModel> Add(RequestAgendamentoViewModel atendimento);
-        Task<ResponseAgendamentoViewModel> Get(int id);
-        Task<List<ResponseAgendamentoViewModel>> List(int idAssociado);
+        Task<ResponseAgendamentoViewModel> Get(long id);
+        Task<List<ResponseAgendamentoViewModel>> List(long idAssociado);
     }
 }
